Extract `createPlanFromJSON` to appActions
